+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 #### Features
 - Save time to database and recall
+- Implement actions after pressing Space or Enter keys
+- Split clock and controls to components
+- Improve interactions with input fields
 
 # 0.1.0 (2020.04.13)